requirements.txt
psycopg2-binary>=2.9.0
torch>=2.0.0
transformers>=4.30.0
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
